Professors and students creation
Adds professors and students

+++ SmartCanteen/models/Client.cs
@@ -9,13 +9,13 @@ namespace SmartCanteen.models
 {
     abstract class Client : User
     {
-        public float Balance  {  get; set; }
+        public decimal Balance  {  get; set; }
 
-        public Client(string name, string nif,float balance) : base(name, nif)
+        public Client(string name, string nif,decimal balance) : base(name, nif)
         {
             string Name = name;
             string Nif = nif;
-            float Balance = balance;
+            decimal Balance = balance;
         }
     }
